worker: move code
Simplify code bu dropping the unnecessary base client package.
